java upgrade to 17 spring version upgrade to 3.0.5 gradle version upgrade to 7.6.1 remove spring cloud package dependency remove sqs package dependency repository implementation move to infrastructure
