+++ domain/build.gradle
@@ -1,5 +1,5 @@
 dependencies {
-    compile "org.springframework.boot:spring-boot-starter-data-jpa:${springbootStarterVersion}"
+    implementation 'jakarta.persistence:jakarta.persistence-api:3.1.0'
 }
 
 jar {
@@ -16,7 +16,7 @@ task testJar(type: Jar, dependsOn: testClasses) {
 }
 
 configurations {
-    tests { extendsFrom testRuntime }
+    tests { extendsFrom testRuntimeOnly }
 }
 
 artifacts {
