+++ domain/src/main/java/com/dolap/product/entity/converter/ProductStatusConverter.java
@@ -2,8 +2,7 @@ package com.dolap.product.entity.converter;
 
 import com.dolap.product.entity.type.ProductStatus;
 
-import javax.persistence.AttributeConverter;
-import javax.persistence.Converter;
+import jakarta.persistence.*;
 
 @Converter
 public class ProductStatusConverter implements AttributeConverter<ProductStatus, Integer> {
