+++ infrastructure/src/main/java/com/dolap/product/rest/interceptor/CorrelationIdInterceptor.java
@@ -1,14 +1,14 @@
 package com.dolap.product.rest.interceptor;
 
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.MDC;
-import org.springframework.web.servlet.handler.HandlerInterceptorAdapter;
+import org.springframework.web.servlet.HandlerInterceptor;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 import java.util.UUID;
 
-public class CorrelationIdInterceptor extends HandlerInterceptorAdapter {
+public class CorrelationIdInterceptor implements HandlerInterceptor {
 
     private static final String CORRELATION_ID_HEADER_NAME = "x-correlation-id";
     private static final String CORRELATION_ID_LOG_VAR_NAME = "correlationId";
