+++ infrastructure/src/test/java/com/dolap/product/rest/controller/base/integration/BaseControllerIntegrationTest.java
@@ -3,15 +3,14 @@ package com.dolap.product.rest.controller.base.integration;
 import com.dolap.product.base.integration.BaseIntegrationTest;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import jakarta.persistence.EntityManager;
+import jakarta.transaction.Transactional;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.transaction.annotation.Transactional;
-
-import javax.persistence.EntityManager;
 
 @AutoConfigureMockMvc
 @Transactional
